+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 - [#1598](https://github.com/clj-kondo/clj-kondo/issues/1598): `:scope-end-row` is missing on multi-arity fn args ([@mainej](https://github.com/mainej))
 - [#1588](https://github.com/clj-kondo/clj-kondo/issues/1588): analyze type hint in reified method
+- [#1581](https://github.com/clj-kondo/clj-kondo/issues/1581): redundant fn wrapper false positive when using pre-post-map
 
 ## 2022.02.09